Add files via upload
It is possible to determine the features which the neural network is identifying in a given spectra X when predicting each stellar parameter. This can be done by calculating the partial derivatives of each of the outputs (Teff, logg, [Fe/H], vsin(i)) with respect to every input flux value of the spectrum and built the network’s Jacobian. For this test, 300 Jacobians were calculated and averaged for 300 randomly chosen spectra. This image, for example, visualize the Jacobian for one of the outputs with respect to all input neurons. From a physical perspective we can clearly see the high dependency on the Hydrogen Balmer lines at 4860A, the FeII line from 4500A to 4530A, and the magneisum triplet at 4480A, and different weak metallic lines.